requirements.txt
pymupdf
anthropic
fastapi
uvicorn[standard]
sqlalchemy
pillow
python-multipart
pydantic
python-dotenv
